backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
networkx==3.3
anthropic==0.28.0
pydantic==2.7.0
python-dotenv==1.0.1
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.7
python-multipart==0.0.9
